# load balancing: Check each host at most once in LeastRequests LB