@euroelessar 2020-04-30
Benchmark before:
```
BM_LeastRequestLoadBalancerChooseHost/100/1/1000000          186 ms        186 ms          4 mean_hits=10k relative_stddev_hits=0.0109756 stddev_hits=109.756
BM_LeastRequestLoadBalancerChooseHost/100/2/1000000          216 ms        216 ms          3 mean_hits=10k relative_stddev_hits=9.08017m stddev_hits=90.8017
BM_LeastRequestLoadBalancerChooseHost/100/3/1000000          243 ms        243 ms          3 mean_hits=10k relative_stddev_hits=9.44174m stddev_hits=94.4174
BM_LeastRequestLoadBalancerChooseHost/100/10/1000000         442 ms        442 ms          2 mean_hits=10k relative_stddev_hits=9.81851m stddev_hits=98.1851
BM_LeastRequestLoadBalancerChooseHost/100/50/1000000        1547 ms       1547 ms          1 mean_hits=10k relative_stddev_hits=9.03159m stddev_hits=90.3159
BM_LeastRequestLoadBalancerChooseHost/100/100/1000000       2928 ms       2928 ms          1 mean_hits=10k relative_stddev_hits=9.8252m stddev_hits=98.252
```
After:
```
BM_LeastRequestLoadBalancerChooseHost/100/1/1000000          199 ms        199 ms          4 mean_hits=10k relative_stddev_hits=0.0106399 stddev_hits=106.399
BM_LeastRequestLoadBalancerChooseHost/100/2/1000000          227 ms        227 ms          3 mean_hits=10k relative_stddev_hits=9.26935m stddev_hits=92.6935
BM_LeastRequestLoadBalancerChooseHost/100/3/1000000          255 ms        255 ms          3 mean_hits=10k relative_stddev_hits=9.63982m stddev_hits=96.3982
BM_LeastRequestLoadBalancerChooseHost/100/10/1000000         449 ms        449 ms          2 mean_hits=10k relative_stddev_hits=9.85124m stddev_hits=98.5124
BM_LeastRequestLoadBalancerChooseHost/100/50/1000000        1571 ms       1571 ms          1 mean_hits=10k relative_stddev_hits=9.47864m stddev_hits=94.7864
BM_LeastRequestLoadBalancerChooseHost/100/100/1000000       2933 ms       2933 ms          1 mean_hits=10k relative_stddev_hits=0.0106751 stddev_hits=106.751
```