@tonya11en 2020-05-12
@euroelessar I think I'm missing something and need a bit of context. Can you clarify the problem this patch is solving? The reason I'm squinting here is that the Azar et. al. paper that introduces P2C was explicit about the choices being independent (in addition to uniformly at random).

The benefit I see here is that it slightly skews the distribution towards hosts with less outstanding requests. Looking at an extreme example of 3 hosts (a, b, c) with requests outstanding looking like (a < b < c), the probabilities for picking each one looks like:

Independent selections
| Outstanding requests | Selection probability |  
|---|---|
| a | ~50% |  
| b  |  ~33% | 
| c  |  ~16% | 

Unique choices
| Outstanding requests | Selection probability |  
|---|---|
| a | ~66% |  
| b  |  ~33% | 
| c  |  0% | 

This doesn't seem to really affect selection probabilities for larger (>5) numbers of hosts in any appreciable way. I ran a quick simulation with 10000 selections below.

Forcing unique selections:
```
0: 1.000000
1: 0.000000
=====
0: 0.674100
1: 0.325900
2: 0.000000
=====
0: 0.504700
1: 0.331700
2: 0.163600
3: 0.000000
=====
0: 0.399500
1: 0.301200
2: 0.203000
3: 0.096300
4: 0.000000
=====
0: 0.328900
1: 0.266500
2: 0.204400
3: 0.136300
4: 0.063900
5: 0.000000
```

Independent selections (the original way):
```
0: 0.751100
1: 0.248900
=====
0: 0.557400
1: 0.336600
2: 0.106000
=====
0: 0.434700
1: 0.307100
2: 0.190800
3: 0.067400
=====
0: 0.367100
1: 0.278900
2: 0.198400
3: 0.115800
4: 0.039800
=====
0: 0.308700
1: 0.245100
2: 0.192900
3: 0.138700
4: 0.086400
5: 0.028200
```

If you do decide to go through with this patch and toggle the unique selections for lower numbers, the simulation results above might help.