@SaveTheRbtz 2020-05-15
@tonya11en nice simulation! Couple of notes below.

1. Seems like your model is a bit too static: in real life once you assign a request to the backend it will have one more active request and less likely to be picked by P2C on the next iteration.

2. The variable we want to optimize is not the backend load itself but some metric around request processing time.  It is totally fine to have an imbalanced load on backends that have different performance characteristics.

The more advanced model would probably have:
* backends with multimodal performance distributions (e.g. backends with different hardware.)  Note that these should have bounded concurrency to reflect their finite performance.
* requests arriving at independent random intervals (in other words requests do not synchronize.)
* requests have log-normally distributed execution time (e.g. backend supports multiple types of APIs, most are very quick, some are quite slow though.) Note that execution time also depends on the chosen backend's performance.