@jmarantz 2020-05-15
This is very cool. Is the simulation in https://blog.envoyproxy.io/examining-load-balancing-algorithms-with-envoy-1be643ea121c in OSS?