@tonya11en 2020-05-15
@jmarantz for the blog post, I made a few changes to [bufferbloater](https://github.com/tonya11en/bufferbloater) for the heterogeneous upstream tests and just reused some code from [my old thesis](https://github.com/tonya11en/thesis/tree/master/code/simulations) for the selection histograms.

My previous comment was just something I cooked up in a Python shell.