@tonya11en 2020-05-15
@SaveTheRbtz I should have clarified that the point in the simulations was to come up with numbers that represent the _probability_ that a request will be sent to a particular node- latency isn't considered of mentioned.You're right in your comments though and echoing a lot of the points made in the blog post @jmarantz is referencing.