@euroelessar 2020-05-17
I have written a small script for simulating different load balancing techniques:
https://github.com/euroelessar/load_balancing/blob/master/load_balancer_simulation.py

It doesn't take "overloaded backends" into account, but does consider that different backends may have different performance. Simulation measure time needed to complete serving all requests, assuming that total concurrency is bounded.

Example of a scenario when ensuring unique choices works noticeably better and is on par with a full scan:
```
$ python3 load_balancer_simulation.py --backends="1,2" --concurrency=2
Name                              Duration    Max Inflight Requests
------------------------------  ----------  -----------------------
least_requests_full_scan              6668                        1
least_requests_unique_2pc             6668                        1
least_requests_unique_4pc             6668                        1
least_requests_1_random               7194                        2
least_requests_independent_2pc        7194                        2
least_requests_independent_4pc        7194                        2
round_robin                           7500                        2
random                                7507                        2

$ python3 load_balancer_simulation.py --backends="1,2" --concurrency=16 
Name                              Duration    Max Inflight Requests
------------------------------  ----------  -----------------------
least_requests_full_scan               834                        8
least_requests_unique_2pc              834                        8
least_requests_unique_4pc              834                        8
least_requests_1_random                860                       13
least_requests_independent_2pc         860                       13
least_requests_independent_4pc         860                       13
random                                 939                       16
round_robin                            939                       12
```